pyproject.toml
[tool.poetry]
name = "word-image-collage"
version = "0.1.0"
description = ""
authors = ["Ziga Vucko <ziga.vucko@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6"
python-dotenv = "0.10.3"
opencv-python = "4.1.1.26"
pillow = "6.2.1"
pyqt5 = "5.13.2"
google-api-python-client = "1.7.11"
python-twitter = "3.5"
wikipedia-api = "0.5.3"
unidecode = "1.1.1"
scipy = "1.3.2"
numpy = "1.17.4"
tqdm = "4.38.0"
scikit-image = "^0.17.2"
PyQt5-stubs = "^5.14"
google-cloud-translate = "2.0.1"
google-cloud-storage="1.38.0"
nltk = "^3.6"
num2words = "^0.5.10"

[build-system]
requires = ["poetry>=0.12"]